+++ trunk/CHANGELOG.rst
@@ -1,4 +1,9 @@
 
+python-stdnet 0.2.2 - 2010 July 7
+========================================
+ * RelatedManager is derived by Manager and therefore implements both all and filter methods.
+ * 10 tests
+
 python-stdnet 0.2.0  - 2010 June 21
 ========================================
  * First official release in pre-alpha